+++ doc/source/conf.py
@@ -133,7 +133,6 @@ html_theme_options = {
     "github_user": "pytest-dev",
     "github_repo": "pytest-testinfra",
     "github_button": True,
-    "travis_button": True,
     "extra_nav_links": {
         "View on github": "https://github.com/pytest-dev/pytest-testinfra",
     },